+++ NEWS.md
@@ -1,5 +1,12 @@
-# austraits 1.0.9
+# austraits 1.0.0
 
 * Improved `load_austraits`
+* Added `get_versions`
+* Added `summarise_trait_means` to aid `trait_pivot_wider`
+* Added `summarise_austraits`
+* Added `lookup_trait`
+* Added `print.austraits`
+* Added `as_wide_table`
 * Improved `get_versions`
 * Added `get_version_latest`
+